added nav buttons & a generator for them
decided to go with calling() syntax over react </> component syntax to
supress key-related errors.  using a call, the top-level object in the
result-array has a key value.  Not having to repeat my key & to values
each time I wanted to add a nav button was part of the point of making
the wrapper, so using </> became of-itself less useful than ()